Basic input and output model support for APIs.
Restored work I did in Dovetail Bootstrap to support emitting
JSON Schema for the output and input models involved in API calls.
I removed this from the master branch as it required replacing Fubu's
IJsonWriter with one using NewtonSoft's JSON serializer which supports JSON Schema.

I am not happy with the JSON Schema being emitted as it is diving deep as Swagger does not.